Script can be used to data from a mame instance for use in testing. [skip ci]
